fix(Coloring): Even after changing the settings, the configuration remains unchanged.
