+++ src/index.ts
@@ -1,33 +1,121 @@
 import "@logseq/libs";
-import { settingUI } from './setting';
 import { logseq as PL } from "../package.json";
 const pluginId = PL.id;
 import Swal from 'sweetalert2';//https://sweetalert2.github.io/
+import CSSmain from './main.css?inline';
+import CSStodayJournal from './todayJournal.css?inline';
+import CSSrainbowJournal from './rainbowJournal.css?inline';
+import CSSadmonitions from './admonition.css?inline';
+import { LSPluginBaseInfo, SettingSchemaDesc } from "@logseq/libs/dist/LSPlugin.user";
 
-//hex -> rgba
-//Credit: https://www.yoheim.net/blog.php?q=20171007
-function hex2rgba(hex, alpha = 1) {
-  //long #FF0000
-  let r = hex.match(/^#([0-9a-f]{2})([0-9a-f]{2})([0-9a-f]{2})$/i);
-  let c = null;
-  if (r) {
-    c = r.slice(1, 4).map(function (x) {
-      return parseInt(x, 16);
-    });
+const keyTagColoring = "tagColoring";
+const keyPageColoring = "pageColoring";
+
+
+//main
+const main = () => {
+  settingUI(); /* -setting */
+  //Logseq bugs fix
+  /* Fix "Extra space when journal queries are not active #6773" */
+  /* background conflict journal queries */
+  /* journal queries No shadow */
+
+  //CSS minify https://csscompressor.com/
+  logseq.provideStyle({ key: "main", style: CSSmain });
+
+  //set today journal coloring
+  const keyTodayJournal = "todayJournal";
+  if (logseq.settings?.todayJournal) {
+    logseq.provideStyle({ key: keyTodayJournal, style: CSStodayJournal });
   }
-  //short #F00
-  r = hex.match(/^#([0-9a-f])([0-9a-f])([0-9a-f])$/i);
-  if (r) {
-    c = r.slice(1, 4).map(function (x) {
-      return 0x11 * parseInt(x, 16);
-    });
+
+  //set rainbow-journal
+  const keyRainbowJournal = "rainbowJournal";
+  if (logseq.settings?.rainbowJournal) {
+    logseq.provideStyle({ key: keyRainbowJournal, style: CSSrainbowJournal });
   }
-  //NG return null
-  if (!c) {
-    return null;
+
+  //set admonitions
+  const keyAdmonitions = "admonitions";
+  if (logseq.settings?.admonitions) {
+    logseq.provideStyle({ key: keyAdmonitions, style: CSSadmonitions });
   }
-  return `rgba(${c[0]}, ${c[1]}, ${c[2]}, ${alpha})`;
-}
+
+  const keyBulletClosed = "bulletClosed";
+  //set BulletClosed
+  const CSSbulletClosed = (color): string => {
+    return `div#app-container span.bullet{height:9px;width:9px;border-radius:40%;opacity:.6;transition:unset!important}
+div#app-container span.bullet-container.bullet-closed{height:11px;width:11px;outline:5px solid ${hex2rgba(color, 0.6)}
+`};
+  logseq.provideStyle({ key: keyBulletClosed, style: CSSbulletClosed(logseq.settings?.bulletClosedColor) });
+
+  //provideColoring(logseq.settings);
+
+
+
+  /* toolbarItem */
+  logseq.App.registerUIItem("toolbar", {
+    key: pluginId,
+    template: `
+  <div data-on-click="open_color_settings" style="font-size:20px">
+  <svg width="20.2" height="20" viewBox="0 0 80 80" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M35.0182 46.589C33.7497 46.2081 32.5148 45.6785 31.3397 45C26.9837 42.4851 23.9873 38.2365 23.0182 33.4109C21.7497 33.7919 20.5148 34.3215 19.3397 35C13.9807 38.094 10.6794 43.812 10.6794 50C10.6794 56.188 13.9807 61.906 19.3397 65C24.6987 68.094 31.3012 68.094 36.6602 65C37.8861 64.2922 39.0043 63.4472 40 62.4906C36.6431 59.2654 34.6794 54.7725 34.6794 50C34.6794 48.8418 34.7951 47.7001 35.0182 46.589Z" fill="#27AE60" />
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M31.3397 15C36.6987 11.906 43.3012 11.906 48.6602 15C54.0192 18.094 57.3205 23.812 57.3205 30C57.3205 31.1582 57.2048 32.2999 56.9817 33.411C52.4654 32.0548 47.5236 32.5845 43.3397 35C42.1138 35.7078 40.9956 36.5529 40 37.5095C39.0043 36.5529 37.8861 35.7078 36.6602 35C32.4763 32.5845 27.5345 32.0548 23.0182 33.411C22.7951 32.2999 22.6794 31.1582 22.6794 30C22.6794 23.812 25.9807 18.094 31.3397 15Z" fill="#EB5757" />
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M44.9817 46.5891C45.2049 47.7002 45.3205 48.8418 45.3205 50C45.3205 54.7725 43.3568 59.2654 40 62.4906C40.9957 63.4472 42.1139 64.2922 43.3397 65C48.6987 68.094 55.3013 68.094 60.6603 65C66.0192 61.906 69.3205 56.188 69.3205 50C69.3205 43.812 66.0192 38.094 60.6603 35C59.4851 34.3215 58.2502 33.7919 56.9817 33.411C56.0127 38.2365 53.0162 42.4851 48.6603 45C47.4851 45.6785 46.2502 46.2081 44.9817 46.5891Z" fill="#2D9CDB" />
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M44.9817 46.5891C40.4655 47.9452 35.5236 47.4156 31.3397 45C26.9838 42.4851 23.9873 38.2365 23.0183 33.411C27.5345 32.0548 32.4764 32.5844 36.6603 35C41.0162 37.5149 44.0127 41.7635 44.9817 46.5891Z" fill="#F2994A" />
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M56.9817 33.411C56.0127 38.2365 53.0162 42.4851 48.6603 45C44.4764 47.4156 39.5345 47.9452 35.0183 46.5891C35.9873 41.7635 38.9838 37.5149 43.3397 35C47.5236 32.5844 52.4655 32.0548 56.9817 33.411Z" fill="#BB6BD9" />
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M40 37.5094C43.3568 40.7346 45.3205 45.2275 45.3205 50C45.3205 54.7725 43.3568 59.2654 40 62.4906C36.6432 59.2654 34.6795 54.7725 34.6795 50C34.6795 45.2275 36.6432 40.7346 40 37.5094Z" fill="#56CCF2" />
+    <path fill-rule="evenodd" clip-rule="evenodd" d="M35.0183 46.5891C35.7146 43.1216 37.4578 39.952 40 37.5095C42.5423 39.952 44.2855 43.1216 44.9818 46.5891C41.734 47.5643 38.2661 47.5643 35.0183 46.5891Z" fill="#F2F2F2" />
+  </svg>
+  </div>
+  `,
+  });
+
+  /* Block slash command */
+  logseq.Editor.registerSlashCommand('🌈Select Admonition panel', async ({ uuid }) => {
+    selectAdmonition(uuid);
+  });
+
+  /* Block ContextMenuItem  */
+  logseq.Editor.registerBlockContextMenuItem('🌈Select Admonition', async ({ uuid }) => {
+    selectAdmonition(uuid);
+  });
+
+  // Setting changed
+  logseq.onSettingsChanged(async (newSet: LSPluginBaseInfo['settings'], oldSet: LSPluginBaseInfo['settings']) => {
+    if (newSet !== oldSet) {
+      await removeProvideStyle(keyTagColoring);
+      await removeProvideStyle(keyPageColoring);
+      provideColoring(newSet);
+      if (oldSet.admonitions !== false && newSet.admonitions === false) {
+        removeProvideStyle(keyAdmonitions);
+      } else if (oldSet.admonitions !== true && newSet.admonitions === true) {
+        logseq.provideStyle({ key: keyAdmonitions, style: CSSadmonitions });
+      }
+      if (oldSet.rainbowJournal !== false && newSet.rainbowJournal === false) {
+        removeProvideStyle(keyRainbowJournal);
+      } else if (oldSet.rainbowJournal !== true && newSet.rainbowJournal === true) {
+        logseq.provideStyle({ key: keyRainbowJournal, style: CSSrainbowJournal });
+      }
+      if (oldSet.todayJournal !== false && newSet.todayJournal === false) {
+        removeProvideStyle(keyTodayJournal);
+      } else if (oldSet.todayJournal !== true && newSet.todayJournal === true) {
+        logseq.provideStyle({ key: keyTodayJournal, style: CSStodayJournal });
+      }
+      await removeProvideStyle(keyBulletClosed);
+      logseq.provideStyle({ key: keyBulletClosed, style: CSSbulletClosed(newSet.bulletClosedColor) });
+    }
+  });
+
+  logseq.provideModel({
+    //toolbar onclick
+    open_color_settings() {
+      logseq.showSettingsUI();
+    }
+  });
+
+};
+//main end
 
 
 //for tag
@@ -65,9 +153,19 @@ const FavoriteOverflowCSS = (page: IPage) => {
   }
 };
 
-//for tag and page
-const refreshSettings = (e) => {
-  const settings = e || {};
+
+const removeProvideStyle = (className: string) => {
+  const doc = parent.document.head.querySelector(`style[data-injected-style^="${className}"]`);
+  if (doc) {
+    doc.remove();
+  }
+};
+
+
+//Tag Coloring & Page Coloring
+const provideColoring = (e) => {
+  const settings = e;
+  if (!settings) { return };
   const settingKeys = Object.keys(settings || {});
   //tag
   const tcArray = settingKeys
@@ -85,7 +183,7 @@ const refreshSettings = (e) => {
     }
   });
   const thisStyle = settingArray.map(generateTagStyle).join("\n");
-  logseq.provideStyle(thisStyle);
+  logseq.provideStyle({ key: keyTagColoring, style: thisStyle });
   //tag end
 
   //page
@@ -104,93 +202,14 @@ const refreshSettings = (e) => {
     }
   });
   const thisStylePage = settingArrayPage.map(generatePageStyle).join("\n");//Page Coloring && favorites coloring
-  settingArrayPage.map(FavoriteOverflowCSS);//favorite recent remove
-  logseq.provideStyle(thisStylePage);
+  //settingArrayPage.map(FavoriteOverflowCSS);//favorite recent remove
+  logseq.provideStyle({ key: keyPageColoring, style: thisStylePage });
   //page end
-
-
-  //set BulletClosed
-  const SettingBulletClosedColor: string = settings.bulletClosedColor;
-  const bulletClosedStyle: string = `div#app-container span.bullet{height:9px;width:9px;border-radius:40%;opacity:.6;transition:unset!important}
-  div#app-container span.bullet-container.bullet-closed{height:11px;width:11px;outline:5px solid ${hex2rgba(SettingBulletClosedColor, 0.6)}
-  `;
-  logseq.provideStyle(bulletClosedStyle);
 };
-
-
-//CSS
-async function ProvideStyle() {
-  //Logseq bugs fix
-  /* Fix "Extra space when journal queries are not active #6773" */
-  /* background conflict journal queries */
-  /* journal queries No shadow */
-  logseq.provideStyle(String.raw`
-  body[data-page="home"] div#today-queries>div.lazy-visibility{min-height:unset!important}body[data-page="home"] div#today-queries>div.lazy-visibility>div.shadow{display:none}body[data-page="home"] div#today-queries div.color-level div.blocks-container,body[data-page="home"] div#today-queries div.color-level{background-color:unset}
-  `);
-  //set today journal coloring
-  if (logseq.settings?.todayJournal) {
-    parent.document.body.classList.add('pc-todayJournal');
-  }
-  logseq.provideStyle(`body[data-page="home"].pc-todayJournal div.light-theme div#journals div.journal-item.content:first-child{border-radius:0.4em;background:rgba(251,255,177,0.2);outline:3px double rgba(251,255,177,0.6);outline-offset:3px}
-  body[data-page="home"].pc-todayJournal div.light-theme div#journals div.journal-item.content:nth-of-type(2){border-radius:0.4em;background:rgba(144,238,144,0.16);outline:3px double rgba(144,238,144,0.6);outline-offset:3px}
-  `);
-
-
-  //set rainbow-journal
-  if (logseq.settings?.rainbowJournal) {
-    parent.document.body.classList.add('pc-rainbowJournal');
-  }
-
-  //set admonitions
-  if (logseq.settings?.admonitions) {
-    parent.document.body.classList.add('pc-admonitions');
-  }
-  logseq.provideStyle(String.raw`
-  body[data-page="page"] div#app-container div.page-blocks-inner,body[data-page="page"] div#app-container div.page-properties,div#app-container :is(h1,h2,h3,h4,h5,h6).with-bg-color,div#app-container div.embed-block,div.relative div.lazy-visibility div.references div.references-blocks div.content>div>div.lazy-visibility,div.references div.references-blocks div.content>div>div.lazy-visibility{outline-offset:2px;outline:2px solid var(--ls-table-tr-even-background-color);border-radius:10px}
-  article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading001"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading002"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading003"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading004"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading005"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading006"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading007"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading008"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading009"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading0010"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading0011"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading0012"]{color:green}
-  article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading101"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading102"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading103"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading104"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading105"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading106"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading107"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading108"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading109"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading1010"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading1011"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="heading1012"]{color:orange}
-  article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn1"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn2"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn3"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn4"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn5"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn6"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn7"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn8"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn9"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn10"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn11"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pn12"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn1"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn2"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn3"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn4"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn5"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn6"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn7"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn8"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn9"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn10"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn11"],article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tn12"]{width: 48%}
-  body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc1"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc2"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc3"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc4"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc5"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc6"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc7"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc8"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc9"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc10"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc11"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc12"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc1"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc2"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc3"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc4"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc5"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc6"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc7"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc8"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc9"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc10"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc11"],body.is-awUi-enabled article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc12"]{left: 48%;margin-top:-4.4em}
-  body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc1"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc2"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc3"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc4"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc5"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc6"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc7"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc8"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc9"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc10"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc11"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="pc12"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc1"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc2"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc3"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc4"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc5"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc6"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc7"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc8"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc9"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc10"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc11"],body:not(.is-awUi-enabled) article div[data-id="logseq-plugin-panel-coloring"].panel-wrap div[data-key="tc12"]{left: 48%;position:absolute;margin-top:-5.4em}
-  body.pc-rainbowJournal div#app-container div.block-children [level="3"]{border-right:4px solid #ff00005c}
-  body.pc-rainbowJournal div#app-container div.block-children [level="4"]{border-right:6px solid #ec87225e}
-  body.pc-rainbowJournal div#app-container div.block-children [level="5"]{border-right:8px solid #ffff0052}
-  body.pc-rainbowJournal div#app-container div.block-children [level="6"]{border-right:10px solid #65d95b7a}
-  body.pc-rainbowJournal div#app-container div.block-children [level="7"]{border-right:12px solid #2ca0df5e}
-  body.pc-rainbowJournal div#app-container div.block-children [level="8"]{border-right:14px solid #9f8af061}
-  body.pc-rainbowJournal div#app-container div.block-children [level="9"]{border-right:16px solid #f15bf74f}
-  body.pc-admonitions main div#app-container a.tag[data-ref*="caution"],body.pc-admonitions main div#app-container a.tag[data-ref*="declaration"]{color:inherit;padding:2px;border-radius:3px;background:rgba(248,180,0,0.7)}
-  body.pc-admonitions main div#app-container div[data-refs-self*='caution'],body.pc-admonitions main div#app-container div[data-refs-self*='declaration']{padding:1.4em;border-radius:16px;background:rgba(248,180,0,0.1);outline:3px double rgba(248,180,0,0.8);outline-offset:3px}
-  body.pc-admonitions main div#app-container div[data-refs-self*='caution']::before,body.pc-admonitions main div#app-container div[data-refs-self*='declaration']::before{content:"🟡";font-size:20px;position:absolute;right:10px;top:10px}
-  body.pc-admonitions div.dark-theme main div#app-container a.tag[data-ref*="caution"],body.pc-admonitions div.dark-theme main div#app-container a.tag[data-ref*="declaration"]{background:rgba(255,248,220,0.3)}
-  body.pc-admonitions div.dark-theme main div#app-container div[data-refs-self*='caution'],body.pc-admonitions div.dark-theme main div#app-container div[data-refs-self*='declaration']{background:rgba(255,248,220,0.1);outline:3px double rgba(255,248,220,0.7)}
-  body.pc-admonitions main div#app-container a.tag[data-ref*="success"],body.pc-admonitions main div#app-container a.tag[data-ref*="facts"]{color:inherit;padding:2px;border-radius:3px;background:rgba(42,178,123,0.6)}
-  body.pc-admonitions main div#app-container div[data-refs-self*='success'],body.pc-admonitions main div#app-container div[data-refs-self*='facts']{padding:1.4em;border-radius:16px;background:rgba(42,178,123,0.1);outline:3px double rgba(42,178,123,0.7);outline-offset:3px}
-  body.pc-admonitions main div#app-container div[data-refs-self*='success']::before,body.pc-admonitions main div#app-container div[data-refs-self*='facts']::before{content:"🟢";font-size:20px;position:absolute;right:10px;top:10px}
-  body.pc-admonitions main div#app-container a.tag[data-ref*="warning"],body.pc-admonitions main div#app-container a.tag[data-ref*="learned"]{color:inherit;padding:2px;border-radius:3px;background:rgba(255,127,80,0.7)}
-  body.pc-admonitions main div#app-container div[data-refs-self*='warning'],body.pc-admonitions main div#app-container div[data-refs-self*='learned']{padding:1.4em;border-radius:16px;background:rgba(255,127,80,0.1);outline:3px double rgba(255,127,80,0.7);outline-offset:3px}
-  body.pc-admonitions main div#app-container div[data-refs-self*='warning']::before,body.pc-admonitions main div#app-container div[data-refs-self*='learned']::before{content:"🟠";font-size:20px;position:absolute;right:10px;top:10px}
-  body.pc-admonitions main div#app-container a.tag[data-ref*="failed"],body.pc-admonitions main div#app-container a.tag[data-ref*="remedy"]{color:inherit;padding:2px;border-radius:3px;background:rgba(220,20,60,0.5)}
-  body.pc-admonitions main div#app-container div[data-refs-self*='failed'],body.pc-admonitions main div#app-container div[data-refs-self*='remedy']{padding:1.4em;border-radius:16px;background:rgba(220,20,60,0.1);outline:3px double rgba(220,20,60,0.7);outline-offset:3px}
-  body.pc-admonitions main div#app-container div[data-refs-self*='failed']::before,body.pc-admonitions main div#app-container div[data-refs-self*='remedy']::before{content:"🔴";font-size:20px;position:absolute;right:10px;top:10px}
-  body.pc-admonitions main div#app-container a.tag[data-ref*="question"],body.pc-admonitions main div#app-container a.tag[data-ref*="discovery"]{color:inherit;padding:2px;border-radius:3px;background:rgba(147,112,219,0.5)}
-  body.pc-admonitions main div#app-container div[data-refs-self*='question'],body.pc-admonitions main div#app-container div[data-refs-self*='discovery']{padding:1.4em;border-radius:16px;background:rgba(147,112,219,0.1);outline:3px double rgba(147,112,219,0.7);outline-offset:3px}
-  body.pc-admonitions main div#app-container div[data-refs-self*='question']::before,body.pc-admonitions main div#app-container div[data-refs-self*='discovery']::before{content:"🟣";font-size:20px;position:absolute;right:10px;top:10px}
-  body.pc-admonitions main div#app-container a.tag[data-ref*="report"],body.pc-admonitions main div#app-container a.tag[data-ref*="note"],body.pc-admonitions main div#app-container a.tag[data-ref*="review"]{color:inherit;padding:2px;border-radius:3px;background:rgba(160,82,45,0.5)}
-  body.pc-admonitions main div#app-container div[data-refs-self*='report'],body.pc-admonitions main div#app-container div[data-refs-self*='note'],body.pc-admonitions main div#app-container div[data-refs-self*='review']{padding:1.4em;border-radius:16px;background:rgba(160,82,45,0.1);outline:3px double rgba(160,82,45,0.7);outline-offset:3px}
-  body.pc-admonitions main div#app-container div[data-refs-self*='report']::before,body.pc-admonitions main div#app-container div[data-refs-self*='note']::before,body.pc-admonitions main div#app-container div[data-refs-self*='review']::before{content:"🟤";font-size:20px;position:absolute;right:10px;top:10px}
-  body.pc-admonitions main div#app-container a.tag[data-ref*="notice"],body.pc-admonitions main div#app-container a.tag[data-ref*="info"],body.pc-admonitions main div#app-container a.tag[data-ref*="review"]{color:inherit;padding:2px;border-radius:3px;background:rgba(30,144,255,0.5)}
-  body.pc-admonitions main div#app-container div[data-refs-self*='notice'],body.pc-admonitions main div#app-container div[data-refs-self*='info'],body.pc-admonitions main div#app-container div[data-refs-self*='review']{padding:1.4em;border-radius:16px;background:rgba(30,144,255,0.1);outline:3px double rgba(30,144,255,0.7);outline-offset:3px}
-  body.pc-admonitions main div#app-container div[data-refs-self*='notice']::before,body.pc-admonitions main div#app-container div[data-refs-self*='info']::before,body.pc-admonitions main div#app-container div[data-refs-self*='review']::before{content:"🔵";font-size:20px;position:absolute;right:10px;top:10px}
-  `);
-
-  console.info(`#${pluginId}: provide style`); /* -plugin-id */
-}
 
 
 //admonition selector
 function selectAdmonition(uuid) {
-  const block = logseq.Editor.getBlock(uuid);
   //dialog
   logseq.showMainUI();
   Swal.fire({
@@ -234,91 +253,134 @@ function selectAdmonition(uuid) {
 }
 
 
-//main
-const main = async () => {
-  console.info(`#${pluginId}: MAIN`); /* -plugin-id */
-  ProvideStyle();
-  refreshSettings(logseq.settings);
-};
-//main end
-
-
-//after
-const after = async () => {
-  console.info(`#${pluginId}: AFTER`); /* -plugin-id */
-
-  settingUI(); /* -setting */
-
-  /* toolbarItem */
-  logseq.App.registerUIItem("toolbar", {
-    key: pluginId,
-    template: `
-  <div data-on-click="open_color_settings" style="font-size:20px">
-  <svg width="20.2" height="20" viewBox="0 0 80 80" fill="none" xmlns="http://www.w3.org/2000/svg">
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M35.0182 46.589C33.7497 46.2081 32.5148 45.6785 31.3397 45C26.9837 42.4851 23.9873 38.2365 23.0182 33.4109C21.7497 33.7919 20.5148 34.3215 19.3397 35C13.9807 38.094 10.6794 43.812 10.6794 50C10.6794 56.188 13.9807 61.906 19.3397 65C24.6987 68.094 31.3012 68.094 36.6602 65C37.8861 64.2922 39.0043 63.4472 40 62.4906C36.6431 59.2654 34.6794 54.7725 34.6794 50C34.6794 48.8418 34.7951 47.7001 35.0182 46.589Z" fill="#27AE60" />
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M31.3397 15C36.6987 11.906 43.3012 11.906 48.6602 15C54.0192 18.094 57.3205 23.812 57.3205 30C57.3205 31.1582 57.2048 32.2999 56.9817 33.411C52.4654 32.0548 47.5236 32.5845 43.3397 35C42.1138 35.7078 40.9956 36.5529 40 37.5095C39.0043 36.5529 37.8861 35.7078 36.6602 35C32.4763 32.5845 27.5345 32.0548 23.0182 33.411C22.7951 32.2999 22.6794 31.1582 22.6794 30C22.6794 23.812 25.9807 18.094 31.3397 15Z" fill="#EB5757" />
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M44.9817 46.5891C45.2049 47.7002 45.3205 48.8418 45.3205 50C45.3205 54.7725 43.3568 59.2654 40 62.4906C40.9957 63.4472 42.1139 64.2922 43.3397 65C48.6987 68.094 55.3013 68.094 60.6603 65C66.0192 61.906 69.3205 56.188 69.3205 50C69.3205 43.812 66.0192 38.094 60.6603 35C59.4851 34.3215 58.2502 33.7919 56.9817 33.411C56.0127 38.2365 53.0162 42.4851 48.6603 45C47.4851 45.6785 46.2502 46.2081 44.9817 46.5891Z" fill="#2D9CDB" />
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M44.9817 46.5891C40.4655 47.9452 35.5236 47.4156 31.3397 45C26.9838 42.4851 23.9873 38.2365 23.0183 33.411C27.5345 32.0548 32.4764 32.5844 36.6603 35C41.0162 37.5149 44.0127 41.7635 44.9817 46.5891Z" fill="#F2994A" />
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M56.9817 33.411C56.0127 38.2365 53.0162 42.4851 48.6603 45C44.4764 47.4156 39.5345 47.9452 35.0183 46.5891C35.9873 41.7635 38.9838 37.5149 43.3397 35C47.5236 32.5844 52.4655 32.0548 56.9817 33.411Z" fill="#BB6BD9" />
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M40 37.5094C43.3568 40.7346 45.3205 45.2275 45.3205 50C45.3205 54.7725 43.3568 59.2654 40 62.4906C36.6432 59.2654 34.6795 54.7725 34.6795 50C34.6795 45.2275 36.6432 40.7346 40 37.5094Z" fill="#56CCF2" />
-    <path fill-rule="evenodd" clip-rule="evenodd" d="M35.0183 46.5891C35.7146 43.1216 37.4578 39.952 40 37.5095C42.5423 39.952 44.2855 43.1216 44.9818 46.5891C41.734 47.5643 38.2661 47.5643 35.0183 46.5891Z" fill="#F2F2F2" />
-  </svg>
-  </div>
-  `,
-  });
-
-  /* Block slash command */
-  logseq.Editor.registerSlashCommand('🌈Select Admonition panel', async ({ uuid }) => {
-    selectAdmonition(uuid);
-  });
-
-  /* Block ContextMenuItem  */
-  logseq.Editor.registerBlockContextMenuItem('🌈Select Admonition', async ({ uuid }) => {
-    selectAdmonition(uuid);
-  });
-
-  parent.document.body.classList.add('is-plugin-panel-coloring-enabled');
-  logseq.beforeunload(async () => {
-    parent.document.body.classList.remove('is-plugin-panel-coloring-enabled');
-  });
-  logseq.onSettingsChanged((settings, oldSet) => {
-    if (settings !== oldSet) {
-      console.log(`#${pluginId}: onSettingsChanged`);
-      refreshSettings(settings);
-      onSettingsChangedCallback(settings, oldSet);
-    }
-  });
-
-  console.info(`#${pluginId}: loaded`);
-};
-
-
-// Setting changed
-const onSettingsChangedCallback = (newSet: any, oldSet: any) => {
-  if (oldSet.admonitions !== false && newSet.admonitions === false) {
-    parent.document.body.classList.remove('pc-admonitions');
-  } else if (oldSet.admonitions !== true && newSet.admonitions === true) {
-    parent.document.body.classList.add('pc-admonitions');
+function hex2rgba(hex: string, alpha: number): string {
+  if (!hex) {
+    throw new Error('Invalid hex color value');
   }
-  if (oldSet.rainbowJournal !== false && newSet.rainbowJournal === false) {
-    parent.document.body.classList.remove('pc-rainbowJournal');
-  } else if (oldSet.rainbowJournal !== true && newSet.rainbowJournal === true) {
-    parent.document.body.classList.add('pc-rainbowJournal');
+  const hexValue = hex.replace('#', '');
+  if (hexValue.length !== 3 && hexValue.length !== 6) {
+    throw new Error('Invalid hex color value');
   }
-  if (oldSet.todayJournal !== false && newSet.todayJournal === false) {
-    parent.document.body.classList.remove('pc-todayJournal');
-  } else if (oldSet.todayJournal !== true && newSet.todayJournal === true) {
-    parent.document.body.classList.add('pc-todayJournal');
-  }
+  const hexArray = hexValue.length === 3 ? hexValue.split('').map(char => char + char) : hexValue.match(/.{2}/g) || [];
+  const rgbaArray = hexArray.map(hexChar => parseInt(hexChar, 16));
+  rgbaArray.push(alpha);
+  return `rgba(${rgbaArray.join(',')})`;
 }
 
 
-const model = {
-  //toolbar onclick
-  open_color_settings() {
-    logseq.showSettingsUI();
-  }
+const settingUI = () => {
+  /* https://logseq.github.io/plugins/types/SettingSchemaDesc.html */
+
+  const rainbowColor = [
+    "#37306B", "#66347F", "#9E4784", "#D27685", "#9DC08B", "#609966", "#40513B", "#060047", "#B3005E", "#E90064", "#FF5F9E", "#E21818"
+  ];
+
+  const generateSettings = () => {
+    const settingArray = [] as SettingSchemaDesc[];
+
+    //option
+    settingArray.push(
+      {
+        key: "admonitions",
+        title: "admonitions by tags",
+        type: "boolean",
+        default: true,
+        description: `
+              🔴#FAILED / #REMEDY, 
+              🟠#WARNING / #LEARNED, 
+              🟡#CAUTION / #DECLARATION, 
+              🟢#SUCCESS / #FACTS, 
+              🔵#NOTICE / #INFO / #REVIEW, 
+              🟣#QUESTION / #DISCOVERY, 
+              🟤#REPORT / #NOTE
+              `,
+      },
+      {
+        key: "rainbowJournal",
+        title: "outline right border",
+        type: "boolean",
+        default: true,
+        description: "Color according to nesting depth.",
+      },
+      {
+        key: "todayJournal",
+        title: "today & yesterday journal coloring",
+        type: "boolean",
+        default: false,
+        description: "background-color: yellow & green (**light theme only)",
+      },
+      {
+        key: `bulletClosedColor`,
+        title: `choice closed-bullet color`,
+        type: "string",
+        default: "f8b400",
+        description: "Accentuate with color",
+        inputAs: "color",
+      },
+    );
+
+
+    //page
+    [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12].forEach((idx) => {
+      settingArray.push(
+        {
+          key: `heading10${idx}`,
+          title: `page [ ${idx} ]`,
+          type: "heading",
+          default: "",
+          description: "Accentuate the specified page like a panel.",
+        },
+        {
+          key: `pn${idx}`,
+          title: `set [ page title ] word`,
+          type: "string",
+          default: ``,
+          description: "",
+        },
+        {
+          key: `pc${idx}`,
+          title: `choice background color`,
+          type: "string",
+          default: rainbowColor[idx - 1],
+          description: "color fades",
+          inputAs: "color",
+        },
+      );
+    });
+
+
+    //tag
+    [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12].forEach((idx) => {
+      settingArray.push(
+        {
+          key: `heading00${idx}`,
+          title: `#tag [ ${idx} ]`,
+          type: "heading",
+          default: "",
+          description: "Accentuate tagged blocks like a panel.",
+        },
+        {
+          key: `tn${idx}`,
+          title: `set [ tag ] word`,
+          type: "string",
+          default: ``,
+          description: "without [ # ]",
+        },
+        {
+          key: `tc${idx}`,
+          title: `choice background color`,
+          type: "string",
+          default: rainbowColor[idx - 1],
+          description: "color fades",
+          inputAs: "color",
+        },
+      );
+    });
+
+    return settingArray;
+  };
+  logseq.useSettingsSchema(generateSettings());
 };
 
 // bootstrap
-logseq.ready(model, main).then(after).catch(console.error);
+logseq.ready(main).catch(console.error);
